+++ Source/Tools/FEXLoader/LinuxSyscalls/ArchHelpers/Arm64_stubs.cpp
@@ -1,4 +1,4 @@
-#include "Interface/Core/ArchHelpers/Arm64.h"
+#include "LinuxSyscalls/ArchHelpers/Arm64.h"
 
 #include <FEXCore/Utils/LogManager.h>
 #include <stdint.h>
